Implement ManeuverMode
This adds maneuvermode to horst. This mode can be toggled over dbus
method and will enable/disable the ADCS rules for detumbling and sunpointing to
allow flying in other configurations.

For safety reasons (e.g. if ADCS mode drains the battery) this
maneuvermode is disabled when entering the safemode and will not
be automatically enabled again.