Retrigger measuring if several signals
If payload daemon is signaling HORST that he wants to measure
we will now ignore the state of the payload daemon and
just send him a doMeasurement again, even if in our state table
this is already running.

This might be neccessary, if we miss a measurement done signal
for some reason (e.g. bug or restart of pl daemon)